+++ Graphs/Tools/grasp_adjacency_gaussian.m
@@ -1,10 +1,13 @@
 %Constructs the adjacency matrix of a graph as a Gaussian kernel of its
-%distance matrix.
+%distance matrix. If the graph already has an adjacency matrix, only the
+%previously defined edges are reweighted.
 %
 %   A = GRASP_ADJACENCY_GAUSSIAN(graph, sigma) construct the adjacency
 %       matrix A such that a_ij = exp(-d_ij² / (2 * sigma²)). Set sigma to
 %       'auto' to use 1/3 of the average distance accross edges of the
-%       graph (heuristic [2]).
+%       graph (heuristic [2]). If the graph already has an adjacency
+%       matrix, only distances corresponding to edges with non-zero weights
+%       are used to compute the mean.
 %
 %   GRASP_ADJACENCY_GAUSSIAN(..., lambda) normalize weights with the power
 %       lambda of each degree prior to computing the Gaussian kernel, i.e.
@@ -19,14 +22,19 @@
 % Authors:
 %  - Benjamin Girault <benjamin.girault@ens-lyon.fr>
 %  - Benjamin Girault <benjamin.girault@usc.edu>
+%  - Benjamin Girault <benjamin.girault@ensai.fr>
 
 % Copyright Benjamin Girault, École Normale Supérieure de Lyon, FRANCE /
 % Inria, FRANCE (2015-11-01)
 % Copyright Benjamin Girault, University of Sourthern California, Los
-% Angeles, California, USA (2017-2019)
+% Angeles, California, USA (2017-2020)
+% Copyright Benjamin Girault, École Nationale de la Statistique et de
+% l'Analyse de l'Information, Bruz, FRANCE (2020-2021)
 % 
 % benjamin.girault@ens-lyon.fr
 % benjamin.girault@usc.edu
+% benjamin.girault@ensai.fr
+
 % 
 % This software is a computer program whose purpose is to provide a Matlab
 % / Octave toolbox for handling and displaying graph signals.
@@ -62,10 +70,21 @@ function A = grasp_adjacency_gaussian(graph, sigma, lambda)
         lambda = 0;
     end
     
+    N = size(graph.distances, 1);
+    
+    has_adja = false;
+    if size(graph.A, 1) == N && size(graph.A, 2) == N
+        has_adja = true;
+        prev_A = (graph.A ~= 0);
+    end
+    
     if ischar(sigma)
         if strcmp(sigma, 'auto')
-            tmp = graph.distances(graph.A ~= 0);
-            sigma = mean(tmp) / 3;
+            tmp = graph.distances;
+            if has_adja
+                tmp = tmp(prev_A);
+            end
+            sigma = mean(tmp(:)) / 3;
         else
             error('Unrecogized paramater ''sigma'': it should be either a postive value or ''auto''.');
         end
@@ -74,10 +93,12 @@ function A = grasp_adjacency_gaussian(graph, sigma, lambda)
         error('Unrecogized paramater ''sigma'': it should be either a postive value or ''auto''.');
     end
     
-    N = size(graph.distances, 1);
     graph.A = exp(-graph.distances .^ 2 / (2 * sigma ^ 2)) - eye(N);
     if lambda ~= 0
         graph = grasp_adjacency_degreenorm(graph, lambda);
     end
     A = graph.A;
+    if has_adja
+        A = A .* prev_A;
+    end
 end